+++ data-raw/preprocess_data.R
@@ -909,6 +909,11 @@ parameters_Thrower1994 <- read.csv("data-raw/Thrower1994_parameters.csv")
 parameters_Thrower1994 <- parameters_Thrower1994 %>%
   select(Species = nfi_species, model_form:source_short)
 
+## Chen & Klinka 2000 (ESSF height growth / site index) ####
+parameters_ChenKlinka1998 <- read.csv("data-raw/ChenKlinka1998_parameters.csv")
+parameters_ChenKlinka1998 <- parameters_ChenKlinka1998 %>%
+  select(Species = nfi_species, b1, b2, b3)
+
 ## Hu & Garcia 2009 (interior spruce site index) ####
 parameters_HuGarcia2009 <- read.csv("data-raw/HuGarcia2009_parameters.csv")
 parameters_HuGarcia2009 <- parameters_HuGarcia2009 %>%
@@ -1386,6 +1391,7 @@ internal_objs <- c(
   "parameters_Nigh2000_gi",
   "parameters_Nigh1998_gi",
   "parameters_Thrower1994",
+  "parameters_ChenKlinka1998",
   "parameters_HuGarcia2009",
   "parameters_Alemdag1991",
   "parameters_Nigh1997",